5-9 centos stream 9 测试-2

+++ include/check_os.sh
@@ -41,46 +41,47 @@ if [ -e "/usr/bin/apt-get" ]; then
   }
 fi
 
-command -v lsb_release >/dev/null 2>&1 || {
-  echo "${CFAILURE}${PM} source failed! ${CEND}"
-  kill -9 $$
-}
+if command -v lsb_release >/dev/null 2>&1; then
+  # Get OS Version
+  OS=$(lsb_release -is)
+  if [[ "${OS}" =~ ^CentOS$|^RedHat$|^Rocky$|^Fedora$|^Amazon$|^Alibaba$|^Aliyun$|^EulerOS$|^openEuler$|^CentOSStream$ ]]; then
+    CentOS_ver=$(lsb_release -rs | awk -F. '{print $1}' | awk '{print $1}')
+    [[ "${OS}" =~ ^Fedora$ ]] && [ ${CentOS_ver} -ge 19 ] >/dev/null 2>&1 && {
+      CentOS_ver=7
+      Fedora_ver=$(lsb_release -rs)
+    }
+    [[ "${OS}" =~ ^Amazon$|^Alibaba$|^Aliyun$|^EulerOS$|^openEuler$ ]] && CentOS_ver=7
+  elif [[ "${OS}" =~ ^Debian$|^Deepin$|^Uos$|^Kali$ ]]; then
+    Debian_ver=$(lsb_release -rs | awk -F. '{print $1}' | awk '{print $1}')
+    [[ "${OS}" =~ ^Deepin$|^Uos$ ]] && [[ "${Debian_ver}" =~ ^20$ ]] && Debian_ver=10
+    [[ "${OS}" =~ ^Kali$ ]] && [[ "${Debian_ver}" =~ ^202 ]] && Debian_ver=10
+  elif [[ "${OS}" =~ ^Ubuntu$|^LinuxMint$|^elementary$ ]]; then
+    Ubuntu_ver=$(lsb_release -rs | awk -F. '{print $1}' | awk '{print $1}')
+    if [[ "${OS}" =~ ^LinuxMint$ ]]; then
+      [[ "${Ubuntu_ver}" =~ ^18$ ]] && Ubuntu_ver=16
+      [[ "${Ubuntu_ver}" =~ ^19$ ]] && Ubuntu_ver=18
+      [[ "${Ubuntu_ver}" =~ ^20$ ]] && Ubuntu_ver=20
+    fi
+    if [[ "${OS}" =~ ^elementary$ ]]; then
+      [[ "${Ubuntu_ver}" =~ ^5$ ]] && Ubuntu_ver=18
+      [[ "${Ubuntu_ver}" =~ ^6$ ]] && Ubuntu_ver=20
+    fi
+  fi
 
-# Get OS Version
-OS=$(lsb_release -is)
-if [[ "${OS}" =~ ^CentOS$|^RedHat$|^Rocky$|^Fedora$|^Amazon$|^Alibaba$|^Aliyun$|^EulerOS$|^openEuler$|^CentOSStream$ ]]; then
-  CentOS_ver=$(lsb_release -rs | awk -F. '{print $1}' | awk '{print $1}')
-  [[ "${OS}" =~ ^Fedora$ ]] && [ ${CentOS_ver} -ge 19 ] >/dev/null 2>&1 && {
-    CentOS_ver=7
-    Fedora_ver=$(lsb_release -rs)
-  }
-  [[ "${OS}" =~ ^Amazon$|^Alibaba$|^Aliyun$|^EulerOS$|^openEuler$ ]] && CentOS_ver=7
-elif [[ "${OS}" =~ ^Debian$|^Deepin$|^Uos$|^Kali$ ]]; then
-  Debian_ver=$(lsb_release -rs | awk -F. '{print $1}' | awk '{print $1}')
-  [[ "${OS}" =~ ^Deepin$|^Uos$ ]] && [[ "${Debian_ver}" =~ ^20$ ]] && Debian_ver=10
-  [[ "${OS}" =~ ^Kali$ ]] && [[ "${Debian_ver}" =~ ^202 ]] && Debian_ver=10
-elif [[ "${OS}" =~ ^Ubuntu$|^LinuxMint$|^elementary$ ]]; then
-  Ubuntu_ver=$(lsb_release -rs | awk -F. '{print $1}' | awk '{print $1}')
-  if [[ "${OS}" =~ ^LinuxMint$ ]]; then
-    [[ "${Ubuntu_ver}" =~ ^18$ ]] && Ubuntu_ver=16
-    [[ "${Ubuntu_ver}" =~ ^19$ ]] && Ubuntu_ver=18
-    [[ "${Ubuntu_ver}" =~ ^20$ ]] && Ubuntu_ver=20
+  # Check OS Version
+  if [ ${CentOS_ver} -lt 6 ] >/dev/null 2>&1 || [ ${Debian_ver} -lt 8 ] >/dev/null 2>&1 || [ ${Ubuntu_ver} -lt 14 ] >/dev/null 2>&1; then
+    echo "${CFAILURE}不支持此系统, 请安装 CentOS 7+,Debian 10+,Ubuntu 18+ ${CEND}"
+    kill -9 $$
   fi
-  if [[ "${OS}" =~ ^elementary$ ]]; then
-    [[ "${Ubuntu_ver}" =~ ^5$ ]] && Ubuntu_ver=18
-    [[ "${Ubuntu_ver}" =~ ^6$ ]] && Ubuntu_ver=20
-  fi
-else
+
+elif [ -e "/etc/redhat-release" ]; then
   # centos stream 9
   OS=$(cat /etc/redhat-release)
   if [ ${OS} == 'CentOS Stream release 9' ]; then
     CentOS_ver=9
     PM=dnf
   fi
-fi
-
-# Check OS Version
-if [ ${CentOS_ver} -lt 6 ] >/dev/null 2>&1 || [ ${Debian_ver} -lt 8 ] >/dev/null 2>&1 || [ ${Ubuntu_ver} -lt 14 ] >/dev/null 2>&1; then
-  echo "${CFAILURE}不支持此系统, 请安装 CentOS 7+,Debian 10+,Ubuntu 18+ ${CEND}"
+else
+  echo "${CFAILURE}${PM} source failed! ${CEND}"
   kill -9 $$
 fi
